Rename MembershipInviteModal and MembershipDeleteModal
