+++ rdmo/core/assets/js/_bs53/components/Modal.js
@@ -2,17 +2,7 @@ import React, { useEffect, useRef } from 'react'
 import PropTypes from 'prop-types'
 import { Modal as BootstrapModal } from 'bootstrap'
 
-const Modal = ({
-  title,
-  show,
-  onClose,
-  onSubmit,
-  submitLabel,
-  submitProps,
-  children,
-  modalProps = {},
-  size = ''
-}) => {
+const Modal = ({ title, show, onClose, onSubmit, submitLabel, submitProps, children, modalProps = {}, size = '' }) => {
   const modalRef = useRef(null)
 
   useEffect(() => {
@@ -31,46 +21,31 @@ const Modal = ({
   }, [show])
 
   return (
-    <div
-      ref={modalRef}
-      className="modal fade"
-      tabIndex="-1"
-      aria-hidden={!show}
-    >
+    <div ref={modalRef} className="modal fade" tabIndex="-1" aria-hidden={!show}>
       <div className={`modal-dialog ${size}`}>
         <div className="modal-content element-modal">
           <div className="modal-header">
             <h2 className="modal-title">{title}</h2>
-            <button
-              type="button"
-              className="btn-close"
-              onClick={onClose}
-              aria-label={gettext('Close')}
-            ></button>
+            <button type="button" className="btn-close" onClick={onClose} aria-label={gettext('Close')}></button>
           </div>
 
-          {children && (
-            <div className="modal-body">
-              {children}
-            </div>
-          )}
+          {
+            children && (
+              <div className="modal-body">
+                {children}
+              </div>
+            )
+          }
 
           <div className="modal-footer">
-            {onSubmit && (
-              <button
-                type="button"
-                className="btn btn-primary"
-                onClick={onSubmit}
-                {...submitProps}
-              >
-                {submitLabel ?? gettext('Save')}
-              </button>
-            )}
-            <button
-              type="button"
-              className="btn btn-secondary"
-              onClick={onClose}
-            >
+            {
+              onSubmit && (
+                <button type="button" className="btn btn-primary" onClick={onSubmit} {...submitProps}>
+                  {submitLabel ?? gettext('Save')}
+                </button>
+              )
+            }
+            <button type="button" className="btn btn-secondary" onClick={onClose}>
               {gettext('Cancel')}
             </button>
           </div>
