--- rdmo/projects/assets/js/project/components/pages/ConfirmDeleteModal.js
@@ -1,84 +1,0 @@
-import React from 'react'
-import PropTypes from 'prop-types'
-import { useDispatch, useSelector } from 'react-redux'
-
-import { Modal } from 'rdmo/core/assets/js/_bs53/components'
-import Html from 'rdmo/core/assets/js/components/Html'
-import { useFieldErrors } from '../../hooks/useFieldErrors'
-import { deleteProjectMember, deleteProjectInvite } from '../../actions/projectActions'
-
-const ConfirmDeleteModal = ({ show, onClose, isManager = false, isMember = false, isCurrentUser = false, person }) => {
-  const dispatch = useDispatch()
-  const { project } = useSelector((state) => state.project.project) ?? {}
-  const errors = useFieldErrors()
-
-  const name =
-    [person.first_name, person.last_name].filter(Boolean).join(' ').trim() ||
-    person.email || ''
-  const text = !isMember ? gettext('Delete invite') : isCurrentUser ? gettext('Leave project') : gettext('Delete membership')
-    return (
-      <Modal
-        title={text}
-        show={show}
-        onClose={onClose}
-        onSubmit={async () => {
-          try {
-            if (isMember) {
-              await dispatch(
-                deleteProjectMember(
-                  person.id,
-                  isCurrentUser && !isManager ? { redirect: true } : {}
-                )
-              )
-            } else {
-              await dispatch(deleteProjectInvite(person.id))
-            }
-            onClose()
-          } catch {
-            // keep modal open on error
-          }
-        }}
-        submitLabel={text}
-        submitProps={{ className: 'btn btn-danger' }}
-        size="modal-md"
-      >
-        <Html
-          html={
-            isCurrentUser
-              ? interpolate(
-                  gettext('You are about to leave the project <b>%s</b>.<br><br>If you want to access this project again, somebody will need to invite you!'),
-                  [project?.title ?? '']
-                )
-              : isMember
-              ? interpolate(
-                  gettext('You are about to remove the user <b>%s</b> from the project <b>%s</b>.'),
-                  [name, project?.title ?? '']
-                )
-              : interpolate(
-                  gettext('You are about to remove the invite of <b>%s</b> from the project <b>%s</b>.'),
-                  [name, project?.title ?? '']
-                )
-          }
-        />
-        {errors.non_field_errors?.map((err, i) => (
-          <div key={i} className="text-danger mt-1">{err}</div>
-        ))}
-      </Modal>
-    )
-}
-
-ConfirmDeleteModal.propTypes = {
-  show: PropTypes.bool.isRequired,
-  onClose: PropTypes.func.isRequired,
-  isManager: PropTypes.bool,
-  isMember: PropTypes.bool,
-  isCurrentUser: PropTypes.bool,
-  person: PropTypes.shape({
-    id: PropTypes.number.isRequired,
-    first_name: PropTypes.string,
-    last_name: PropTypes.string,
-    email: PropTypes.string,
-  })
-}
-
-export default ConfirmDeleteModal
